Split large watcher.go file into more specific watchers.
This commit just moves some code to keep the watcher.go file from
growing. No changes are made.
